apps/orchestrator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ecossistema-orchestrator"
version = "0.1.0"
description = "FastAPI Orchestrator — expõe Anthropic Managed Agents via HTTP/SSE"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "anthropic>=0.40.0",
    "sse-starlette>=2.1.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "supabase>=2.7.0",
    "pyyaml>=6.0.1",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "pytest-mock>=3.14.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/orchestrator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.metadata]
allow-direct-references = true
